Added many license plate numbers
Sources:
https://lihkg.com/thread/1400555/page/1
https://lihkg.com/thread/1372349/page/1

This is merely a collection of online information.
I am not responsible for the accuracy or correctness
of this data. Use at your own risk. In particular,
all expressed or implied warranties, including, but
not limited to, the warranties of merchantibility,
fitness for a particular purpose, and noninfringement,
are expressly disclaimed.